chore(integrations/mail): fix package name
