+++ .run/EmailExampleApplication.run.xml
@@ -2,7 +2,7 @@
   <configuration default="false" name="EmailExampleApplication" type="SpringBootApplicationConfigurationType" factoryName="Spring Boot" nameIsGenerated="true">
     <option name="ACTIVE_PROFILES" value="local" />
     <module name="refarch-email-integration-example" />
-    <option name="SPRING_BOOT_MAIN_CLASS" value="de.muenchen.oss.refarch.email.integration.EmailExampleApplication" />
+    <option name="SPRING_BOOT_MAIN_CLASS" value="de.muenchen.oss.refarch.integration.email.EmailExampleApplication" />
     <extension name="net.ashald.envfile">
       <option name="IS_ENABLED" value="false" />
       <option name="IS_SUBST" value="false" />
