--- refarch-integrations/refarch-email-integration/refarch-email-integration-core/src/main/java/de/muenchen/oss/refarch/email/integration/domain/exception/SendMailException.java
@@ -1,8 +1,0 @@
-package de.muenchen.oss.refarch.email.integration.domain.exception;
-
-@SuppressWarnings("PMD.MissingSerialVersionUID")
-public class SendMailException extends RuntimeException {
-    public SendMailException(final String message, final Throwable cause) {
-        super(message, cause);
-    }
-}
