--- refarch-integrations/refarch-email-integration/refarch-email-integration-core/src/main/java/de/muenchen/oss/refarch/email/integration/domain/exception/TemplateException.java
@@ -1,8 +1,0 @@
-package de.muenchen.oss.refarch.email.integration.domain.exception;
-
-@SuppressWarnings("PMD.MissingSerialVersionUID")
-public class TemplateException extends RuntimeException {
-    public TemplateException(final String message, final Exception cause) {
-        super(message, cause);
-    }
-}
