--- refarch-integrations/refarch-email-integration/refarch-email-integration-core/src/main/java/de/muenchen/oss/refarch/email/integration/domain/model/Attachment.java
@@ -1,8 +1,0 @@
-package de.muenchen.oss.refarch.email.integration.domain.model;
-
-import jakarta.activation.DataSource;
-
-public record Attachment(
-        String fileName,
-        DataSource file) {
-}
