+++ refarch-integrations/refarch-email-integration/refarch-email-integration-core/src/main/java/de/muenchen/oss/refarch/integration/email/domain/exception/SendMailException.java
@@ -1,0 +1,8 @@
+package de.muenchen.oss.refarch.integration.email.domain.exception;
+
+@SuppressWarnings("PMD.MissingSerialVersionUID")
+public class SendMailException extends RuntimeException {
+    public SendMailException(final String message, final Throwable cause) {
+        super(message, cause);
+    }
+}
