+++ refarch-integrations/refarch-email-integration/refarch-email-integration-core/src/main/java/de/muenchen/oss/refarch/integration/email/domain/model/Attachment.java
@@ -1,0 +1,8 @@
+package de.muenchen.oss.refarch.integration.email.domain.model;
+
+import jakarta.activation.DataSource;
+
+public record Attachment(
+        String fileName,
+        DataSource file) {
+}
