+++ refarch-integrations/refarch-email-integration/refarch-email-integration-core/src/main/java/de/muenchen/oss/refarch/integration/email/domain/model/TemplateMail.java
@@ -1,0 +1,31 @@
+package de.muenchen.oss.refarch.integration.email.domain.model;
+
+import jakarta.validation.constraints.NotBlank;
+import jakarta.validation.constraints.NotEmpty;
+import java.util.List;
+import java.util.Map;
+import lombok.EqualsAndHashCode;
+import lombok.Getter;
+
+@EqualsAndHashCode(callSuper = true)
+@Getter
+public class TemplateMail extends BasicMail {
+
+    /**
+     * Template of the mail.
+     */
+    @NotBlank(message = "No template given") private final String template;
+
+    /**
+     * Bottom body of the mail.
+     */
+    @NotEmpty(message = "No content given") private final Map<String, Object> content;
+
+    public TemplateMail(final String receivers, final String receiversCc, final String receiversBcc, final String subject, final String replyTo,
+            final List<Attachment> attachments,
+            final String template, final Map<String, Object> content) {
+        super(receivers, receiversCc, receiversBcc, subject, replyTo, attachments);
+        this.template = template;
+        this.content = content;
+    }
+}
