+++ refarch-integrations/refarch-email-integration/refarch-email-integration-example/src/main/java/de/muenchen/oss/refarch/integration/email/EmailExampleApplication.java
@@ -1,0 +1,26 @@
+package de.muenchen.oss.refarch.integration.email;
+
+import lombok.RequiredArgsConstructor;
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.boot.context.event.ApplicationReadyEvent;
+import org.springframework.context.ApplicationContext;
+import org.springframework.context.event.EventListener;
+
+@SpringBootApplication
+@RequiredArgsConstructor
+public class EmailExampleApplication {
+    private final ExampleMailService exampleMailService;
+    private final ApplicationContext context;
+
+    public static void main(final String[] args) {
+        SpringApplication.run(EmailExampleApplication.class, args);
+    }
+
+    @EventListener(ApplicationReadyEvent.class)
+    public void sendTestMail() {
+        this.exampleMailService.testSendMail();
+        this.exampleMailService.testSendMailTemplate();
+        SpringApplication.exit(context);
+    }
+}
