+++ refarch-integrations/refarch-email-integration/refarch-email-integration-starter/src/main/resources/META-INF/spring/org.springframework.boot.autoconfigure.AutoConfiguration.imports
@@ -1,1 +1,1 @@
-de.muenchen.oss.refarch.email.integration.configuration.MailAutoConfiguration
+de.muenchen.oss.refarch.integration.email.configuration.MailAutoConfiguration
